convert Dashboard to typescript
